Move HunkDiff into UI package
