+++ apps/web/package.json
@@ -22,9 +22,9 @@
 		"@sveltejs/adapter-auto": "^3.0.0",
 		"@sveltejs/kit": "catalog:svelte",
 		"@sveltejs/vite-plugin-svelte": "catalog:svelte",
-		"@types/diff-match-patch": "^1.0.36",
 		"@types/node": "^22.3.0",
 		"@types/rails__actioncable": "^6.1.11",
+		"@types/diff-match-patch": "^1.0.36",
 		"diff-match-patch": "^1.0.5",
 		"svelte": "catalog:svelte",
 		"svelte-check": "catalog:svelte",
