+++ packages/ui/package.json
@@ -65,7 +65,27 @@
 		"vitest": "catalog:",
 		"svelte-french-toast": "^1.2.0",
 		"marked": "^10.0.0",
-		"@types/marked": "^5.0.2"
+		"@types/marked": "^5.0.2",
+		"@codemirror/lang-cpp": "^6.0.2",
+		"@codemirror/lang-css": "^6.2.1",
+		"@codemirror/lang-html": "^6.4.9",
+		"@codemirror/lang-java": "^6.0.1",
+		"@codemirror/lang-javascript": "^6.2.2",
+		"@codemirror/lang-json": "^6.0.1",
+		"@codemirror/lang-markdown": "^6.2.5",
+		"@codemirror/lang-php": "^6.0.1",
+		"@codemirror/lang-python": "^6.1.6",
+		"@codemirror/lang-rust": "^6.0.1",
+		"@codemirror/lang-vue": "^0.1.3",
+		"@codemirror/lang-wast": "^6.0.2",
+		"@codemirror/lang-xml": "^6.1.0",
+		"@codemirror/language": "^6.10.2",
+		"@codemirror/legacy-modes": "^6.4.0",
+		"@lezer/common": "^1.2.1",
+		"@lezer/highlight": "^1.2.0",
+		"@replit/codemirror-lang-svelte": "^6.0.0",
+		"@types/diff-match-patch": "^1.0.36",
+		"diff-match-patch": "^1.0.5"
 	},
 	"peerDependencies": {
 		"svelte": "^4.0.0 || ^5.0.0-0"
